Id of wxCheckListBox is not a long but an int.
git-svn-id: https://svn.code.sf.net/p/wired/code/branches/saveXml@1265 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ src/engine/AudioEngine.h
@@ -145,7 +145,7 @@ static int	AudioCallback(const void *input,
 
 
   int nchan = 0;
-  vector<long>::iterator chan;
+  vector<int>::iterator chan;
 
   if (data->SampleFormat & paFloat32)
     {
